+++ CHANGELOG.md
@@ -23,9 +23,7 @@
   The help output from various commands has greatly improved as well.
   A few notable breaking changes:
     * `doc --spm-module` was removed (use `--spm --module` instead)
-    * `sourcekitten module_info` is now `sourcekitten module-info`
-    * some commands that previously took compiler arguments after a `--`
-      delimiter should now omit the `--` delimiter.  
+    * `sourcekitten module_info` is now `sourcekitten module-info`  
   [JP Simard](https://github.com/jpsim)
 
 ##### Enhancements